pyproject.toml
[project]
name = "mcp-for-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.10.6",
    "infisicalsdk>=1.0.9",
    "openai>=1.99.5",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.8",
]
